pyproject.toml
[tool.poetry]
name = "scrapegraph-test"
version = "0.1.0"
description = ""
authors = ["momonong <morrischen2.9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scrapegraphai = "^1.37.1"
langchain = "^0.3.17"
langchain-community = "^0.3.16"
playwright = "^1.49.1"
pymongo = "^4.11.1"
msal = "^1.31.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
gspread = "^6.1.4"
google-auth = "^2.38.0"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
